feat: Created basic structure for dependency finder model -> not working yet, just prototype
